requirement.txt
streamlit
python-chess
stockfish